Fix integration tests by removing @5 files altogether
