+++ IntegrationTests/Package.swift
@@ -11,7 +11,7 @@ let package = Package(
     .executable(name: "Failing", targets: ["Failing"]),
   ],
   targets: [
-    .target(name: "Spectre", exclude: ["Spectre/XCTest@5.swift"]),
+    .target(name: "Spectre"),
     .target(name: "Passing", dependencies: ["Spectre"]),
     .target(name: "Disabled", dependencies: ["Spectre"]),
     .target(name: "Failing", dependencies: ["Spectre"]),
