+++ Sources/Spectre/XCTest.swift
@@ -29,4 +29,15 @@ class XcodeReporter: ContextReporter {
   func addSuccess(_ name: String)  {}
 
   func addDisabled(_ name: String) {}
+
+  func addFailure(_ name: String, failure: FailureType) {
+    // Xcode 12 removed `recordFailure` and replaced with `record(_:)`
+#if compiler(>=5.3) && os(macOS)
+    let location = XCTSourceCodeLocation(filePath: failure.file, lineNumber: failure.line)
+    let issue = XCTIssue(type: .assertionFailure, compactDescription: "\(name): \(failure.reason)", detailedDescription: nil, sourceCodeContext: .init(location: location), associatedError: nil, attachments: [])
+    testCase.record(issue)
+#else
+    testCase.recordFailure(withDescription: "\(name): \(failure.reason)", inFile: failure.file, atLine: failure.line, expected: false)
+#endif
+  }
 }
